POLIO-127 refactor duplicated code
