+++ plugins/polio/js/src/pages/IM/utils.js
@@ -2,54 +2,7 @@ import React from 'react';
 import CheckIcon from '@material-ui/icons/Check';
 import { Box } from '@material-ui/core';
 import { IM_PASS, IM_FAIL, IM_WARNING } from './constants';
-
 import MESSAGES from '../../constants/messages';
-
-// TODO dedupe as its' the same for LQAS and IM
-const convertRoundDataToArray = roundDataAsDict => {
-    const districtNames = Object.keys(roundDataAsDict);
-    const roundData = Object.values(roundDataAsDict);
-    return roundData.map((value, index) => {
-        return { ...value, name: districtNames[index] };
-    });
-};
-
-// TODO dedupe as its' the same for LQAS and IM
-export const convertLQASData = LQASData => {
-    if (!LQASData) return {};
-    const { stats } = LQASData;
-    const campaignKeys = Object.keys(stats);
-    const result = {};
-    campaignKeys.forEach(key => {
-        if (stats[key]) {
-            result[key] = {};
-            result[key].round_1 = convertRoundDataToArray(stats[key].round_1);
-            result[key].round_2 = convertRoundDataToArray(stats[key].round_2);
-        }
-    });
-    return result;
-};
-// TODO dedupe as its' the same for LQAS and IM
-export const findLQASDataForShape = ({ shape, LQASData, round, campaign }) => {
-    if (!LQASData || !LQASData[campaign]) return null;
-    const dataForRound = LQASData[campaign][round];
-    const result = dataForRound.filter(data => data.district === shape.id)[0];
-    return result;
-};
-// TODO dedupe as its' the same for LQAS and IM
-export const findLQASDataForDistrict = ({
-    district,
-    LQASData,
-    round,
-    campaign,
-}) => {
-    if (!LQASData) return null;
-    const dataForRound = LQASData[campaign][round];
-    const result = dataForRound.filter(
-        data => data.district === district.district,
-    );
-    return result[0];
-};
 
 export const determineStatusForDistrict = district => {
     if (!district) return null;
@@ -60,69 +13,6 @@ export const determineStatusForDistrict = district => {
     return IM_FAIL;
 };
 
-// TODO have exhaustive sorting function
-const sortCampaignNames = (nameA, nameB) => {
-    const [countryCodeA, referenceA] = nameA?.label.split('-');
-    const [countryCodeB, referenceB] = nameB?.label.split('-');
-    const comparison = countryCodeA.localeCompare(countryCodeB, undefined, {
-        sensitivity: 'accent',
-    });
-    if (comparison === 0) {
-        const refA = parseInt(referenceA, 10);
-        const refB = parseInt(referenceB, 10);
-        if (refA < refB) return -1;
-        if (refA > refB) return 1;
-        return 0;
-    }
-    return comparison;
-};
-
-export const makeCampaignsDropDown = campaigns =>
-    campaigns
-        .map(campaign => {
-            return {
-                label: campaign.obr_name,
-                value: campaign.obr_name,
-            };
-        })
-        .sort(sortCampaignNames);
-
-export const defaultShapeStyle = {
-    color: 'grey',
-    opacity: '1',
-    fillColor: 'lightGrey',
-    weight: '1',
-    zIndex: 1,
-};
-
-export const getScopeStyle = (shape, scope) => {
-    const isShapeInScope =
-        scope.filter(shapeInScope => shape.id === shapeInScope.id).length === 1;
-    if (isShapeInScope) {
-        return {
-            color: 'grey',
-            opacity: '1',
-            fillColor: 'grey',
-            weight: '2',
-            zIndex: 1,
-        };
-    }
-    return defaultShapeStyle;
-};
-
-export const findScope = (obrName, campaigns, shapes) => {
-    let scopeIds = [];
-    if (obrName) {
-        scopeIds = campaigns
-            .filter(campaign => campaign.obr_name === obrName)
-            .map(campaign => campaign.group.org_units)
-            .flat();
-    } else {
-        scopeIds = campaigns.map(campaign => campaign.group.org_units).flat();
-    }
-    return shapes.filter(shape => scopeIds.includes(shape.id));
-};
-
 const applyStatusColor = status => {
     if (status === IM_PASS) return { color: 'green' };
     if (status === IM_FAIL) return { color: 'red' };
@@ -130,7 +20,20 @@ const applyStatusColor = status => {
     return null;
 };
 
-export const lqasTableColumns = formatMessage => {
+export const getImStatsForRound = (lqasData, campaign, round) => {
+    if (!lqasData[campaign]) return [[], [], [], []];
+    const totalEvaluated = [...lqasData[campaign][round]];
+    const allStatuses = totalEvaluated.map(district => {
+        return determineStatusForDistrict(district);
+    });
+    const passed = allStatuses.filter(status => status === IM_PASS);
+    const disqualified = allStatuses.filter(status => status === IM_WARNING);
+    const failed = allStatuses.filter(status => status === IM_FAIL);
+
+    return [totalEvaluated, passed, failed, disqualified];
+};
+
+export const imTableColumns = formatMessage => {
     return [
         {
             Header: formatMessage(MESSAGES.districtName),
@@ -172,31 +75,4 @@ export const lqasTableColumns = formatMessage => {
             },
         },
     ];
-};
-
-export const sortDistrictsByName = districts => {
-    return districts.sort((districtA, districtB) =>
-        districtA.name.localeCompare(districtB.name, undefined, {
-            sensitivity: 'accent',
-        }),
-    );
-};
-
-export const getLqasStatsForRound = (lqasData, campaign, round) => {
-    if (!lqasData[campaign]) return [[], [], [], []];
-    const totalEvaluated = [...lqasData[campaign][round]];
-    const allStatuses = totalEvaluated.map(district => {
-        return determineStatusForDistrict(district);
-    });
-    const passed = allStatuses.filter(status => status === IM_PASS);
-    const disqualified = allStatuses.filter(status => status === IM_WARNING);
-    const failed = allStatuses.filter(status => status === IM_FAIL);
-
-    return [totalEvaluated, passed, failed, disqualified];
-};
-
-export const findCountryIds = LqasData => {
-    const { stats } = LqasData;
-    const campaignKeys = Object.keys(stats);
-    return campaignKeys.map(campaignKey => stats[campaignKey].country_id);
 };
